Included orbit.json file. Got replay working through sockets.
Need to try to make it work by passing the data from parent to child.  Can't seem to get that to work.  Also have to implement rooms in socket.

To use the orbit.json file, just use mongoimport in MongoDB and you can use it.

Also want to create a REST API to get the data instead of directly with sockets.